pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.9",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.3",
    "langchain-groq>=1.1.2",
    "pymupdf>=1.27.2.3",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
]
